Move checker function
